refactor: clean up duplicate perp components and reorganize file structure
- Remove duplicate PerpAccountPanel.tsx and SizeInput.tsx from TradingPanel directory
- Delete unused swapAnalytics.ts atom file from MarketDetailV2
- Move PerpAccountDebugInfo component to panels/PerpAccountPanel.tsx
- Fix import paths to use panels subdirectory structure
- Clean up import order in LeverageAdjustModal.tsx
